feat(user): restrict user's delete and update mutations

+++ server/v2/src/entities/user/mutations/userDelete.ts
@@ -5,6 +5,7 @@ import {
 } from 'graphql'
 import { UserModel } from '../userModel'
 import { fromGlobalId, mutationWithClientMutationId } from 'graphql-relay'
+import { getHeadersPayload } from '../../../auth/getHeadersPayload'
 
 export const userDelete = mutationWithClientMutationId({
   name: 'userDelete',
@@ -18,13 +19,36 @@ export const userDelete = mutationWithClientMutationId({
     deletedCount: {
       type: GraphQLString,
       resolve: response => response.deletedCount
+    },
+    error: {
+      type: GraphQLString,
+      resolve: response => response.error
     }
   },
-  mutateAndGetPayload: async ({ id }) => {
-    const response = await UserModel.deleteOne({
-      _id: fromGlobalId(id).id
-    })
+  mutateAndGetPayload: async ({ id }, ctx) => {
+    const { error, payload } = getHeadersPayload(ctx)
 
-    return response
+    if (error || payload === null) {
+      return {
+        error: 'Unauthorized',
+        review: null
+      }
+    }
+
+    if (payload.id !== fromGlobalId(id).id) {
+      return {
+        error: 'Unauthorized'
+      }
+    }
+
+    try {
+      const response = await UserModel.deleteOne({
+        _id: fromGlobalId(id).id
+      })
+
+      return response
+    } catch (error: any) {
+      return { error: error.message }
+    }
   }
 })
